feat(vmm): EVENT_IDX notification suppression for net RX (ABX-351)
Extract shared VirtIO queue helpers (read_used_idx, write_used_entry,
should_notify) from blk_worker.rs into new virtqueue_util.rs module.

In net_rx_worker, gate interrupt injection on should_notify() —
VirtIO spec 2.7.7.2 EVENT_IDX check. The guest's virtio-net NAPI
driver writes used_event to indicate when it wants the next
interrupt. The net-io thread skips set_spi + hv_vcpus_exit when
the guest hasn't consumed up to used_event yet.

VIRTIO_F_EVENT_IDX (bit 29) was already negotiated (net.rs:617).
No feature bit or FDT changes needed.

Expected: 10-50x fewer interrupts during sustained transfer.
